+++ lib/InvertedClipper.dart
@@ -5,13 +5,14 @@ import 'WidgetData.dart';
 
 class InvertedClipper extends CustomClipper<Path> {
 
-  final Animation<double> _animation;
+  final Animation<double> animation;
   final List<WidgetData> widgetsData;
+  double padding;
   Function deepEq = const DeepCollectionEquality().equals;
   List<HoleArea> areas = [];
 
-  InvertedClipper(this._animation, Listenable reclip,
-      {this.widgetsData}) : super(reclip: reclip) {
+  InvertedClipper({@required this.padding, this.animation, Listenable reclip,
+      this.widgetsData}) : super(reclip: reclip) {
     if (widgetsData.isNotEmpty) {
       widgetsData.forEach((WidgetData widgetData) {
         if (widgetData.isEnabled) {
@@ -31,21 +32,22 @@ class InvertedClipper extends CustomClipper<Path> {
   @override
   Path getClip(Size size) {
     Path path = Path();
+    double animationValue = animation != null ? animation.value : 0;
     areas.forEach((HoleArea area) {
       switch (area.shape) {
         case WidgetShape.Oval: {
-          path.addOval(Rect.fromLTWH(area.x - ((area.padding + _animation.value*15) / 2), area.y - (area.padding + _animation.value*15) / 2,
-              area.width + (area.padding + _animation.value*15), area.height + (area.padding + _animation.value*15)));
+          path.addOval(Rect.fromLTWH(area.x - (((area.padding + padding) + animationValue*15) / 2), area.y - ((area.padding + padding) + animationValue*15) / 2,
+              area.width + ((area.padding + padding) + animationValue*15), area.height + ((area.padding + padding) + animationValue*15)));
         }
         break;
         case WidgetShape.Rect: {
-          path.addRect(Rect.fromLTWH(area.x - ((area.padding + _animation.value*15) / 2), area.y - (area.padding + _animation.value*15) / 2,
-              area.width + (area.padding + _animation.value*15), area.height + (area.padding + _animation.value*15)));
+          path.addRect(Rect.fromLTWH(area.x - (((area.padding + padding) + animationValue*15) / 2), area.y - ((area.padding + padding) + animationValue*15) / 2,
+              area.width + ((area.padding + padding) + animationValue*15), area.height + ((area.padding + padding) + animationValue*15)));
         }
         break;
         case WidgetShape.RRect: {
-          path.addRRect(RRect.fromRectAndCorners(Rect.fromLTWH(area.x - ((area.padding + _animation.value*15) / 2), area.y - (area.padding + _animation.value*15) / 2,
-              area.width + (area.padding + _animation.value*15), area.height + (area.padding + _animation.value*15)),
+          path.addRRect(RRect.fromRectAndCorners(Rect.fromLTWH(area.x - (((area.padding + padding) + animationValue*15) / 2), area.y - ((area.padding + padding) + animationValue*15) / 2,
+              area.width + ((area.padding + padding) + animationValue*15), area.height + ((area.padding + padding) + animationValue*15)),
               topLeft: Radius.circular(5.0),
               topRight: Radius.circular(5.0),
               bottomLeft: Radius.circular(5.0),